[Dagger Refactor]: Move ComponentGenerator logic into its own sub-package.
New: Added an export of jsr250 library, which is needed for the @Generated annotation of Dagger's generated sources.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=323408703
