+++ java/dagger/internal/codegen/SourceFileGeneratorsModule.java
@@ -16,12 +16,9 @@
 
 package dagger.internal.codegen;
 
-import dagger.Binds;
 import dagger.Module;
 import dagger.Provides;
 import dagger.internal.codegen.base.SourceFileGenerator;
-import dagger.internal.codegen.binding.BindingGraph;
-import dagger.internal.codegen.binding.ComponentDescriptor;
 import dagger.internal.codegen.binding.MembersInjectionBinding;
 import dagger.internal.codegen.binding.ProductionBinding;
 import dagger.internal.codegen.binding.ProvisionBinding;
@@ -54,20 +51,6 @@ abstract class SourceFileGeneratorsModule {
       MembersInjectorGenerator generator, CompilerOptions compilerOptions) {
     return hjarWrapper(generator, compilerOptions);
   }
-
-  @Binds
-  abstract SourceFileGenerator<BindingGraph> componentGenerator(ComponentGenerator generator);
-
-  // The HjarSourceFileGenerator wrapper first generates the entire TypeSpec before stripping out
-  // things that aren't needed for the hjar. However, this can be really expensive for the component
-  // because it is usually the most expensive file to generate, and most of its content is not
-  // needed in the hjar. Thus, instead of wrapping the ComponentGenerator in HjarSourceFileGenerator
-  // we provide a completely separate processing step, ComponentHjarProcessingStep, and generator,
-  // ComponentHjarGenerator, for when generating hjars for components, which can avoid generating
-  // the parts of the component that would have been stripped out by the HjarSourceFileGenerator.
-  @Binds
-  abstract SourceFileGenerator<ComponentDescriptor> componentHjarGenerator(
-      ComponentHjarGenerator hjarGenerator);
 
   @Provides
   @ModuleGenerator
