+++ java/dagger/internal/codegen/componentgenerator/CurrentImplementationSubcomponent.java
@@ -14,13 +14,13 @@
  * limitations under the License.
  */
 
-package dagger.internal.codegen;
+package dagger.internal.codegen.componentgenerator;
 
 import dagger.BindsInstance;
 import dagger.Subcomponent;
-import dagger.internal.codegen.ComponentImplementationBuilder.RootComponentImplementationBuilder;
-import dagger.internal.codegen.ComponentImplementationBuilder.SubcomponentImplementationBuilder;
 import dagger.internal.codegen.binding.BindingGraph;
+import dagger.internal.codegen.componentgenerator.ComponentImplementationBuilder.RootComponentImplementationBuilder;
+import dagger.internal.codegen.componentgenerator.ComponentImplementationBuilder.SubcomponentImplementationBuilder;
 import dagger.internal.codegen.writing.ComponentBindingExpressions;
 import dagger.internal.codegen.writing.ComponentImplementation;
 import dagger.internal.codegen.writing.ComponentRequirementExpressions;
@@ -35,11 +35,13 @@ import java.util.Optional;
  */
 @Subcomponent
 @PerComponentImplementation
-interface CurrentImplementationSubcomponent {
+// This only needs to be public because the type is referenced by generated component.
+public interface CurrentImplementationSubcomponent {
   RootComponentImplementationBuilder rootComponentBuilder();
 
   SubcomponentImplementationBuilder subcomponentBuilder();
 
+  /** Returns the builder for {@link CurrentImplementationSubcomponent}. */
   @Subcomponent.Builder
   interface Builder {
     @BindsInstance
